requirements_lang_mac.txt
agentops==0.3.26
aixplain-internal==0.2.6rc0
celery==5.4.0
coverage==7.5.3
datasets==3.5.0
docker==7.1.0
docx==0.2.4
exceptiongroup==1.2.2
fastapi==0.115.6
fqdn==1.5.1
future==1.0.0
google_search_results==2.4.2
httpx-sse==0.4.0
huggingface==0.0.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipywidgets==8.1.6
isoduration==20.11.0
langchain-experimental==0.0.65
langchain-openai==0.1.8
langchainhub==0.1.18
langgraph-checkpoint-sqlite==2.0.1
langgraph-cli==0.1.59
langgraph-sdk==0.1.42
langserve==0.2.2
loguru==0.7.3
networkx==3.4.2
nltk==3.9.1
notebook==7.2.2
ollama==0.4.7
openpyxl==3.1.5
pathvalidate==3.2.3
pickleshare==0.7.5
pip==24.2
plotly==5.24.1
pre_commit==4.0.1
proglog==0.1.11
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
pymed==0.8.9
pypdf==5.4.0
PyPDF2==3.0.1
pytesseract==0.3.13
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-runner==6.0.0
python-docx==1.1.2
python-pptx==1.0.2
redis==5.2.0
requests-mock==1.12.1
scikit-learn==1.5.2
seaborn==0.13.2
sentry-sdk==2.24.1
serpapi==0.1.5
sqlparse==0.5.3
sse-starlette==2.1.0
statsmodels==0.14.4
sympy==1.13.3
tavily-python==0.5.0
trustcall==0.0.26
uri-template==1.3.0
webcolors==24.11.1
wikipedia==1.4.0
wordcloud==1.9.4
